wont load as well as much missing code that needs to be written trying without help i've corrected some errors but don't think it's correct
