+++ src/main/java/com/thewheatking/minecraftfarmertechmod/common/blockentity/machines/EnergyControllerBlockEntity.java
@@ -1,0 +1,32 @@
+package com.thewheatking.minecraftfarmertechmod.common.blockentity.machines;
+
+import com.thewheatking.minecraftfarmertechmod.common.blockentity.base.BaseMachineBlockEntity;
+
+import static com.thewheatking.minecraftfarmertechmod.hybrid.HybridBlockEntities.ENERGY_CONTROLLER;
+
+public class EnergyControllerBlockEntity extends BaseMachineBlockEntity {
+    public EnergyControllerBlockEntity(net.minecraft.core.BlockPos pos, net.minecraft.world.level.block.state.BlockState state) {
+        super(ENERGY_CONTROLLER.get(), pos, state, 100000, 5000, 5000, 27);
+        // Implementation needed
+    }
+
+    @Override
+    protected boolean canOperate() {
+        return false;
+    }
+
+    @Override
+    protected void performOperation() {
+
+    }
+
+    @Override
+    public net.minecraft.network.chat.Component getDisplayName() {
+        return net.minecraft.network.chat.Component.translatable("block.minecraftfarmertechmod.energy_controller");
+    }
+
+    @Override
+    public net.minecraft.world.inventory.AbstractContainerMenu createMenu(int containerId, net.minecraft.world.entity.player.Inventory playerInventory, net.minecraft.world.entity.player.Player player) {
+        return null; // TODO: Implement menu
+    }
+}
