+++ src/main/java/com/thewheatking/minecraftfarmertechmod/common/blockentity/machines/NetworkBridgeBlockEntity.java
@@ -1,0 +1,32 @@
+package com.thewheatking.minecraftfarmertechmod.common.blockentity.machines;
+
+import com.thewheatking.minecraftfarmertechmod.common.blockentity.base.BaseMachineBlockEntity;
+
+import static com.thewheatking.minecraftfarmertechmod.hybrid.HybridBlockEntities.NETWORK_BRIDGE;
+
+public class NetworkBridgeBlockEntity extends BaseMachineBlockEntity {
+    public NetworkBridgeBlockEntity(net.minecraft.core.BlockPos pos, net.minecraft.world.level.block.state.BlockState state) {
+        super(NETWORK_BRIDGE.get(), pos, state, 100000, 20000, 20000, 27);
+        // Implementation needed
+    }
+
+    @Override
+    protected boolean canOperate() {
+        return false;
+    }
+
+    @Override
+    protected void performOperation() {
+
+    }
+
+    @Override
+    public net.minecraft.network.chat.Component getDisplayName() {
+        return net.minecraft.network.chat.Component.translatable("block.minecraftfarmertechmod.network_bridge");
+    }
+
+    @Override
+    public net.minecraft.world.inventory.AbstractContainerMenu createMenu(int containerId, net.minecraft.world.entity.player.Inventory playerInventory, net.minecraft.world.entity.player.Player player) {
+        return null; // TODO: Implement menu
+    }
+}
